==> 3. RST-LSTM.ipynb <==
### Train the RST-LSTM

from tqdm import tqdm
import pandas as pd
import os
import json
import pickle

for fold in tqdm(range(5), desc="Writing train/test files for mask-related classification"):
    pathname = f'data/fold_{fold}'
    
    train = pickle.load(open(f'data/fold_{fold}/train.pkl', 'rb'))
    train['tree'] = train['annot'].map(lambda row: row['rst'][0])
    train['label1'] = train.masks_stance.map(str)
    train['label2'] = train.masks_argument.map(str)
    train[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/train_mask_rst.pkl')

    test = pickle.load(open(f'data/fold_{fold}/test.pkl', 'rb'))
    test['tree'] = test['annot'].map(lambda row: row['rst'][0])
    test['label1'] = test.masks_stance.map(str)
    test['label2'] = test.masks_argument.map(str)
    test[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/test_mask_rst.pkl')

print(test.iloc[0].tree)

#### Config for masks 

%%writefile configs/rstbert_masks_0.jsonnet

local embedding_dim = 768;
local foldnum = 0;
local max_length = 512;
local lr = std.parseJson(std.extVar('lr'));
local dropout = std.parseJson(std.extVar('dropout'));
local batch_size = std.parseJson(std.extVar('batch_size'));
local treelstm_hidden_size = std.parseJson(std.extVar('treelstm_hidden_size'));
local dataset_reader_type = "models_scripts.dataset_readers.rst_tree_reader.IsaNLPRSTDatasetReader";
local model_type = "models_scripts.models.two_outputs_tree_lstm_clf.TwoOutputsTreeLSTM";
local model_name = "DeepPavlov/rubert-base-cased-conversational";

{
  vocabulary: {
    non_padded_namespaces: ["tokens", "labels1", "labels2"]
  },
  dataset_reader: {
      type: dataset_reader_type,
      num_labels1: 4,
      num_labels2: 4,
      tokenizer: {
        type: "pretrained_transformer",
        model_name: model_name,
        max_length: max_length,
      },
      token_indexers: {
          tokens: {
            type: "pretrained_transformer",
            model_name: model_name,
            max_length: max_length,
            namespace: 'tokens',
      },
    },
  },
  train_data_path: 'data/fold_' + foldnum + '/train_mask_rst.pkl',
  validation_data_path: 'data/fold_' + foldnum + '/test_mask_rst.pkl',
  model: {
    type: model_type,
    dropout: dropout,
    num_labels1: 4,
    num_labels2: 4,
    label1_weights: [0.2, 0.3, 1.0, 1.0],
    label2_weights: [0.1, 0.1, 1.0, 1.0],
    treelstm_hidden_size: treelstm_hidden_size,
    text_field_embedder: {
        token_embedders: {
            tokens: {
              type: "pretrained_transformer",
              model_name: model_name,
              max_length: max_length,
            },
        },
    },
    seq2vec_encoder: {
        type: "bert_pooler",
           pretrained_model: model_name,
           requires_grad: false,
           dropout: 0.2,
    },
    initializer: {
        regexes: [
        ["_text_field_embedder.*|_seq2vec_encoder.*|_classification_layer1.*|_classification_layer2.*",
            {
                "type": "pretrained",
                "weights_file_path": "convbert_masks/fold_" + foldnum + "/weights.th",
            }
        ],
        ],
    },
  },
  data_loader: {
    batch_sampler: {
      type: 'bucket',
      batch_size: batch_size,
    },
  },
  trainer: {
    optimizer: {
      type: "adam",
      lr: lr,
    },
    learning_rate_scheduler: {
      type: "slanted_triangular",
      cut_frac: 0.06
    },
    grad_norm: 5.0,
    grad_clipping: 5.0,
    validation_metric: '+all_mean',
    num_serialized_models_to_keep: 1,
    num_epochs: 20,
    patience: 3,
    cuda_device: 0,
  },
}

%%writefile configs/hparams.json

[
  {
    "type": "int",
    "attributes": {
      "name": "treelstm_hidden_size",
      "low": 50,
      "high": 100
    }
  },
  {
    "type": "int",
    "attributes": {
      "name": "batch_size",
      "low": 1,
      "high": 12
    }
  },
  {
    "type": "float",
    "attributes": {
      "name": "dropout",
      "low": 0.3,
      "high": 0.5
    }
  },
  {
    "type": "float",
    "attributes": {
      "name": "lr",
      "low": 2e-5,
      "high": 2e-3,
      "log": true
    }
  }
]

%%writefile tune_rstbert_masks.sh

export METHOD=rstbert_masks
rm -r $METHOD
mkdir $METHOD
export FOLD=0
export STUDY_NAME=rstbert_test1
allennlp tune configs/${METHOD}_${FOLD}.jsonnet configs/hparams.json --serialization-dir $METHOD/fold_${FOLD} \
    --study-name $STUDY_NAME \
    --skip-if-exists \
    --metrics best_validation_all_mean \
    --direction maximize
rm rstbert_masks/fold_0/trial_*/*.th

import glob
import json
import os


def find_top_trials(path):
    mean_all = dict()
    for directory in glob.glob(path):
        try:
            metrics = json.load(open(os.path.join(directory, 'metrics.json'), 'r'))
            mean_all[directory] = (metrics.get('best_validation_all_mean'), 
                                   metrics.get('best_validation_f1_1_mean'),
                                   metrics.get('best_validation_f1_2_mean'))
        except:
            pass
    return {key: value for key, value in sorted(mean_all.items(), key=lambda x: x[1][0], reverse=True)}

ma = find_top_trials('rstbert_masks/fold_0/trial_*')
ma

! allennlp best-params --study-name rstbert_test1

%%writefile rstbert_masks.sh

export METHOD=rstbert_masks
rm -r $METHOD
mkdir $METHOD
export batch_size=4
export dropout=0.4
export lr=3e-05
export treelstm_hidden_size=94

python utils/make_k_copies.py --filename configs/${METHOD}_0.jsonnet --k 5

export FOLD=0
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_mask_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_mask_rst.pkl \
#                  --include-package models_scripts

export FOLD=1
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_mask_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_mask_rst.pkl \
#                  --include-package models_scripts

export FOLD=2
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_mask_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_mask_rst.pkl \
#                  --include-package models_scripts

export FOLD=3
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_mask_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_mask_rst.pkl \
#                  --include-package models_scripts

export FOLD=4
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_mask_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_mask_rst.pkl \
#                  --include-package models_scripts

#### Config for vaccines 

for fold in tqdm(range(5), desc="Writing train/test files for vaccines-related classification"):
    pathname = f'data/fold_{fold}'
    
    train = pickle.load(open(f'data/fold_{fold}/train.pkl', 'rb'))
    train['tree'] = train['annot'].map(lambda row: row['rst'][0])
    train['label1'] = train.vaccines_stance.map(str)
    train['label2'] = train.vaccines_argument.map(str)
    train[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/train_vaccines_rst.pkl')

    test = pickle.load(open(f'data/fold_{fold}/test.pkl', 'rb'))
    test['tree'] = test['annot'].map(lambda row: row['rst'][0])
    test['label1'] = test.vaccines_stance.map(str)
    test['label2'] = test.vaccines_argument.map(str)
    test[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/test_vaccines_rst.pkl')

%%writefile configs/rstbert_vaccines_0.jsonnet

local embedding_dim = 768;
local foldnum = 0;
local max_length = 512;
local lr = std.parseJson(std.extVar('lr'));
local dropout = std.parseJson(std.extVar('dropout'));
local batch_size = std.parseJson(std.extVar('batch_size'));
local treelstm_hidden_size = std.parseJson(std.extVar('treelstm_hidden_size'));
local dataset_reader_type = "models_scripts.dataset_readers.rst_tree_reader.IsaNLPRSTDatasetReader";
local model_type = "models_scripts.models.two_outputs_tree_lstm_clf.TwoOutputsTreeLSTM";
local model_name = "DeepPavlov/rubert-base-cased-conversational";

{
  vocabulary: {
    non_padded_namespaces: ["tokens", "labels1", "labels2"]
  },
  dataset_reader: {
      type: dataset_reader_type,
      num_labels1: 4,
      num_labels2: 4,
      tokenizer: {
        type: "pretrained_transformer",
        model_name: model_name,
        max_length: max_length,
      },
      token_indexers: {
          tokens: {
            type: "pretrained_transformer",
            model_name: model_name,
            max_length: max_length,
            namespace: 'tokens',
      },
    },
  },
  train_data_path: 'data/fold_' + foldnum + '/train_vaccines_rst.pkl',
  validation_data_path: 'data/fold_' + foldnum + '/test_vaccines_rst.pkl',
  model: {
    type: model_type,
    dropout: dropout,
    num_labels1: 4,
    num_labels2: 4,
    label1_weights: [0.1, 0.4, 1.0, 1.0],
    label2_weights: [0.1, 0.1, 0.6, 1.0],
    treelstm_hidden_size: treelstm_hidden_size,
    text_field_embedder: {
        token_embedders: {
            tokens: {
              type: "pretrained_transformer",
              model_name: model_name,
              max_length: max_length,
            },
        },
    },
    seq2vec_encoder: {
        type: "bert_pooler",
           pretrained_model: model_name,
           requires_grad: false,
           dropout: 0.2,
    },
    initializer: {
        regexes: [
        ["_text_field_embedder.*|_seq2vec_encoder.*|_classification_layer1.*|_classification_layer2.*",
            {
                "type": "pretrained",
                "weights_file_path": "convbert_vaccines/fold_" + foldnum + "/weights.th",
            }
        ],
        ],
    },
  },
  data_loader: {
    batch_sampler: {
      type: 'bucket',
      batch_size: batch_size,
    },
  },
  trainer: {
    optimizer: {
      type: "adam",
      lr: lr,
#       weight_decay: 0.1,
    },
    learning_rate_scheduler: {
      type: "slanted_triangular",
      cut_frac: 0.06
    },
    grad_norm: 5.0,
    grad_clipping: 5.0,
    validation_metric: '+all_mean',
    num_serialized_models_to_keep: 1,
    num_epochs: 20,
    patience: 3,
    cuda_device: 1,
  },
}

%%writefile tune_rstbert_vaccines.sh

export METHOD=rstbert_vaccines
rm -r $METHOD
mkdir $METHOD
export FOLD=0
export STUDY_NAME=rstbert_test2
# optuna delete-study $STUDY_NAME
allennlp tune configs/${METHOD}_${FOLD}.jsonnet configs/hparams.json --serialization-dir $METHOD/fold_${FOLD} \
    --study-name $STUDY_NAME \
    --skip-if-exists \
    --metrics best_validation_all_mean \
    --direction maximize
rm rstbert_vaccines/fold_0/trial_*/*.th

! allennlp best-params --study-name rstbert_test2

find_top_trials('rstbert_vaccines/fold_0/trial_*')

%%writefile rstbert_vaccines.sh

export METHOD=rstbert_vaccines
rm -r $METHOD
mkdir $METHOD
export batch_size=3
export dropout=0.2
export lr=2.5e-05
export treelstm_hidden_size=53

python utils/make_k_copies.py --filename configs/${METHOD}_0.jsonnet --k 5

export FOLD=0
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_vaccines_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_vaccines_rst.pkl \
#                  --include-package models_scripts

export FOLD=1
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_vaccines_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_vaccines_rst.pkl \
#                  --include-package models_scripts

export FOLD=2
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_vaccines_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_vaccines_rst.pkl \
#                  --include-package models_scripts

export FOLD=3
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_vaccines_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_vaccines_rst.pkl \
#                  --include-package models_scripts

export FOLD=4
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_vaccines_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_vaccines_rst.pkl \
#                  --include-package models_scripts

#### Config for quarantine

for fold in tqdm(range(5), desc="Writing train/test files for quarantine-related classification"):
    pathname = f'data/fold_{fold}'
    
    train = pickle.load(open(f'data/fold_{fold}/train.pkl', 'rb'))
    train['tree'] = train['annot'].map(lambda row: row['rst'][0])
    train['label1'] = train.quarantine_stance.map(str)
    train['label2'] = train.quarantine_argument.map(str)
    train[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/train_quarantine_rst.pkl')

    test = pickle.load(open(f'data/fold_{fold}/test.pkl', 'rb'))
    test['tree'] = test['annot'].map(lambda row: row['rst'][0])
    test['label1'] = test.quarantine_stance.map(str)
    test['label2'] = test.quarantine_argument.map(str)
    test[['text_id', 'tree', 'label1', 'label2']].to_pickle(f'data/fold_{fold}/test_quarantine_rst.pkl')

train.quarantine_stance.map(str).value_counts()

train.quarantine_argument.map(str).value_counts()

%%writefile configs/rstbert_quarantine_0.jsonnet

local embedding_dim = 768;
local foldnum = 0;
local max_length = 512;
local lr = std.parseJson(std.extVar('lr'));
local dropout = std.parseJson(std.extVar('dropout'));
local batch_size = std.parseJson(std.extVar('batch_size'));
local treelstm_hidden_size = std.parseJson(std.extVar('treelstm_hidden_size'));
local dataset_reader_type = "models_scripts.dataset_readers.rst_tree_reader.IsaNLPRSTDatasetReader";
local model_type = "models_scripts.models.two_outputs_tree_lstm_clf.TwoOutputsTreeLSTM";
local model_name = "DeepPavlov/rubert-base-cased-conversational";

{
  vocabulary: {
    non_padded_namespaces: ["tokens", "labels1", "labels2"]
  },
  dataset_reader: {
      type: dataset_reader_type,
      num_labels1: 4,
      num_labels2: 4,
      tokenizer: {
        type: "pretrained_transformer",
        model_name: model_name,
        max_length: max_length,
      },
      token_indexers: {
          tokens: {
            type: "pretrained_transformer",
            model_name: model_name,
            max_length: max_length,
            namespace: 'tokens',
      },
    },
  },
  train_data_path: 'data/fold_' + foldnum + '/train_quarantine_rst.pkl',
  validation_data_path: 'data/fold_' + foldnum + '/test_quarantine_rst.pkl',
  model: {
    type: model_type,
    dropout: dropout,
    num_labels1: 4,
    num_labels2: 4,
    label1_weights: [0.1, 0.1, 0.3, 1.0],
    label2_weights: [0.1, 0.1, 0.6, 1.0],
    treelstm_hidden_size: treelstm_hidden_size,
    text_field_embedder: {
        token_embedders: {
            tokens: {
              type: "pretrained_transformer",
              model_name: model_name,
              max_length: max_length,
            },
        },
    },
    seq2vec_encoder: {
        type: "bert_pooler",
           pretrained_model: model_name,
           requires_grad: false,
           dropout: 0.2,
    },
    initializer: {
        regexes: [
        ["_text_field_embedder.*|_seq2vec_encoder.*|_classification_layer1.*|_classification_layer2.*",
            {
                "type": "pretrained",
                "weights_file_path": "convbert_quarantine/fold_" + foldnum + "/weights.th",
            }
        ],
        ],
    },
  },
  data_loader: {
    batch_sampler: {
      type: 'bucket',
      batch_size: batch_size,
    },
  },
  trainer: {
    optimizer: {
      type: "adam",
      lr: lr,
    },
    learning_rate_scheduler: {
      type: "slanted_triangular",
      cut_frac: 0.06
    },
    validation_metric: '+all_mean',
    num_serialized_models_to_keep: 1,
    num_epochs: 20,
    patience: 3,
    cuda_device: 1,
  },
}

! ls convbert_quarantine/fold_0

%%writefile tune_rstbert_quarantine.sh

export METHOD=rstbert_quarantine
rm -r $METHOD
mkdir $METHOD
export FOLD=0
export STUDY_NAME=rstbert_test31

allennlp tune configs/${METHOD}_${FOLD}.jsonnet configs/hparams.json --serialization-dir $METHOD/fold_${FOLD} \
    --study-name $STUDY_NAME \
    --skip-if-exists \
    --metrics best_validation_all_mean \
    --direction maximize
rm rstbert_quarantine/fold_0/trial_*/*.th

find_top_trials('rstbert_quarantine/fold_0/trial_*')

! allennlp best-params --study-name rstbert_test31

%%writefile rstbert_quarantine.sh

export METHOD=rstbert_quarantine
rm -r $METHOD
mkdir $METHOD
export batch_size=7
export dropout=0.2
export lr=2e-05
export treelstm_hidden_size=124

python utils/make_k_copies.py --filename configs/${METHOD}_0.jsonnet --k 5

export FOLD=0
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_quarantine_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_quarantine_rst.pkl \
#                  --include-package models_scripts

export FOLD=1
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_quarantine_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_quarantine_rst.pkl \
#                  --include-package models_scripts

export FOLD=2
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_quarantine_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_quarantine_rst.pkl \
#                  --include-package models_scripts

export FOLD=3
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_quarantine_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_quarantine_rst.pkl \
#                  --include-package models_scripts

export FOLD=4
allennlp train -s $METHOD/fold_${FOLD} configs/${METHOD}_${FOLD}.jsonnet
allennlp predict --use-dataset-reader --silent \
                 --output-file $METHOD/fold_${FOLD}/predictions_test.json $METHOD/fold_${FOLD}/model.tar.gz \
                 data/fold_${FOLD}/test_quarantine_rst.pkl \
                 --include-package models_scripts
                 
# allennlp predict --use-dataset-reader --silent \
#                  --output-file $METHOD/fold_${FOLD}/predictions_train.json $METHOD/fold_${FOLD}/model.tar.gz \
#                  data/fold_${FOLD}/train_quarantine_rst.pkl \
#                  --include-package models_scripts

### (Optional) A place to prep a model 

import torch

! cd rstbert_masks/fold_0/trial_19/ && tar -xzvf model.tar.gz

model = torch.load('rstbert_masks/fold_0/trial_19/weights.th', map_location='cpu')

relations = ['attribution_N', 'attribution_S',
             'background_N', 'background_S',
             'cause-effect_N', 'cause-effect_S',
             'comparison_N',
             'concession_N', 'concession_S',
             'condition_N', 'condition_S',
             'contrast_N',
             'elaboration_N', 'elaboration_S',
             'evidence_N', 'evidence_S',
             'interpretation-evaluation_N', 'interpretation-evaluation_S',
             'joint_N',
             'preparation_S', 'preparation_N',
             'purpose_N', 'purpose_S',
             'restatement_N',
             'same-unit_N',
             'sequence_N',
             'solutionhood_S', 'solutionhood_N',
             'root'
             ]
weights = dict(zip(relations, model['_treelstm_encoder.W_iou.weight'][:,8:].sum(axis=0).detach().numpy()))

{key: value for key, value in sorted(weights.items(), key=lambda x: x[1], reverse=True)}
